% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_functions.R
\name{C_Clayton}
\alias{C_Clayton}
\title{Clayton copula}
\usage{
C_Clayton(
  x,
  y,
  theta = VineCopula::BiCopTau2Par(family = 3, tau = 0.5, check.taus = TRUE),
  tau = NA
)
}
\arguments{
\item{x}{It is required that \eqn{0 \le x \le 1}}

\item{y}{It is required that \eqn{0 \le y \le 1}}

\item{theta}{Parameter of the Clayton Copula.
It is required that \code{theta} \eqn{\ge -1}.}

\item{tau}{Kendall's \eqn{\tau} for the Clayton Copula. If set to \code{NA}, the
value of Kendall's \eqn{\tau} is defined by \code{theta}.
\itemize{
\item \code{NA}: If no value for \code{tau} is passed to the function,
the parameter \code{theta} will be used.
\item \code{tau}: If a value for \code{tau} is assigned, the parameter \code{theta}
will be chosen such that the passed value for the Kendall's \eqn{\tau} is achieved.
}}
}
\description{
This function evaluates the Clayton Copula distribution at (\code{x}, \code{y}). One can
define the parameter \code{theta} of the Clayton Copula. Alternatively one can
set the target value for Kendall's \eqn{\tau} and the corresponding value of \code{theta}
will be set automatically.
}
\examples{
x <- runif(10^2)
y <- runif(10^2)
sample <- C_Clayton(x, y)
}
