% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_functions.R
\name{C_Gamma_Frailty}
\alias{C_Gamma_Frailty}
\title{Gamma frailty Copula}
\usage{
C_Gamma_Frailty(x, y, alpha = 3)
}
\arguments{
\item{x}{It is required that \eqn{0 \le x \le 1}}

\item{y}{It is required that \eqn{0 \le y \le 1}}

\item{alpha}{Parameter of the Gamma Frailty Copula.
It is required that \code{alpha} \eqn{\ge 1}.}
}
\description{
Gamma frailty Copula
}
\examples{
x <- runif(10^2)
y <- runif(10^2)
sample <- C_Gamma_Frailty(x, y)
}
