% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_functions.R
\name{C_independence}
\alias{C_independence}
\title{Independence Copula}
\usage{
C_independence(x, y)
}
\arguments{
\item{x}{It is required that \eqn{0 \le x \le 1}}

\item{y}{It is required that \eqn{0 \le y \le 1}}
}
\description{
Independence Copula
}
\examples{
x <- runif(10^2)
y <- runif(10^2)
sample <- C_independence(x, y)
}
