% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa - classes.R
\docType{class}
\name{copula_graphic-class}
\alias{copula_graphic-class}
\title{A S4 class containing the copula graphic object}
\description{
A S4 class containing the copula graphic object
}
\section{Slots}{

\describe{
\item{\code{sample_data}}{}
}}

