% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Data Input: load competing risk data}
\usage{
load_data(
  data_path = system.file("inst/Melanoma.csv", package = "copulagraphicr")
)
}
\arguments{
\item{data_path}{Path to file. Files with extension '.csv' are supported.}
}
\description{
The \code{load_data} function loads data from \code{data_path} and extracts
the competing risk data consisting of observations (T\eqn{_i}, \eqn{\delta_i}).
}
\details{
The competing risk data should consist of the observed sample data
(T\eqn{_i}, \eqn{\delta_i}).
\itemize{
\item \verb{competing risks}:
\itemize{
\item \eqn{X} (time to death)
\item \eqn{Y} (censoring time)
}
\item \verb{sample data}:
\itemize{
\item \verb{time to event} \eqn{T =} min(X,Y)
\item \code{event-indicator} \eqn{\delta} = 1 if \eqn{X \le Y}; \eqn{\delta} = 0 if \eqn{X > Y}
}
}
}
\examples{
# load data set: df <- load_data()
}
