% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{Melanoma}
\alias{Melanoma}
\title{Melanoma data}
\description{
The study observed a total of 205 patients during the time from 1962 until 1977.
At the end 134 were still alive, while 57 had died from cancer and 14 died from other causes.
The original dataset contains 12 variables of which we will only consider the following
\itemize{
\item \code{time}: time-to-event in days from operation (event: "status change"),
\item \code{status}: numeric with values (0 = censored, 1 = death.melanoma, 2 = death other).
}
Note that the Copula-Graphic estimator is designed for the case of two competing risks.
Hence, only patients with status 0 or 1 are considered in the Melanoma package data.
The two competing events are A = "died from Melanoma" and B = "censored". The sample data
contains the respective time-to-event \eqn{T} and the information if A or B was observed, captured by \eqn{\delta}.
}
\examples{
# library(riskRegression)
# df <- load(Melanoma)
}
\references{
riskRegression::Melanoma
}
\keyword{data}
