% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{visualize_data}
\alias{visualize_data}
\title{Visualization of competing risk data}
\usage{
visualize_data(sample_data)
}
\arguments{
\item{sample_data}{Sample data consisting of observed competing risk data (T\eqn{_i}, \eqn{\delta_i}).
The dataset should be loaded using the \code{load_data} function.}
}
\description{
Visualization of competing risk data
}
\examples{
\donttest{## load data set:
# data_path <- system.file("inst/Melanoma.csv", package = "copulagraphicr")
# sample_data <- copulagraphicr::load_data(data_path = data_path)
# copulagraphicr::visualize_data(sample_data = sample_data)
}
}
