% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_quantities.R
\name{estimate_quantities}
\alias{estimate_quantities}
\title{Estimate empirical quantities \code{k}, \code{p_1} and \code{p_2}}
\usage{
estimate_quantities(sample_data)
}
\arguments{
\item{sample_data}{Sample data consisting of observed competing risk data (T\eqn{_i}, \eqn{\delta_i}).
The dataset should be loaded using the \code{load_data} function.}
}
\description{
In order to calculate the Copula-Graphic estimator the following quantities
are needed
\deqn{k(t) = pr(X>t, Y>t),
p_1(t) = pr(X \le t, \delta = 1) and p_2(t) = pr(Y \le t, \delta = 0).}
These quantities can be estimated by their empirical estimates
\itemize{
\item \eqn{est k(t_i) = 1/n \sum 1_(X_j > t_i, Y_j > t_i)},
\item \eqn{est p_1(t_i) = 1/n \sum 1_(X_j \le t_i, X_j \le Y_j)},
\item \eqn{est p_2(t_i) = 1/n \sum 1_(X_j \le t_i, X_j > Y_j)}.
}
}
\examples{
\donttest{# load data set:
data_path <- system.file("inst/Melanoma.csv", package = "copulagraphicr")
df <- copulagraphicr::load_data(data_path = data_path)

# estimate empirical quantities:
est <- copulagraphicr::estimate_quantities(sample_data = df)
t_grid <- est[[1]]
k <- est[[2]]
p_1 <- est[[3]]
p_2 <- est[[4]]
}
}
